refactor(math): split TrajectoryMath.CatmullRomFit into its own file
